role_default: Prevent linting of ignore_errors line
Sadly, Ansible still has no support for warnings with common taks and
a debug task is not enough to inform users that they might need an
updated version of the role (or submit a patch upstream, adding a
supported plattform). Failing is also not an option as it might
prevent executing a role on a platform it might work / one knows
what is doing.

So

- ansible.builtin.fail
  msg: "Warning"
  gnore_errors: true # noqa: ignore-errors

still seems to be the best alternative.