role_default: use internal variable for platform compatibility check
This replaces the use of galaxy['platforms'] with a dedicated variable in
vars/main.yml. It is simply more consistent, especially when introducing a
default run role in the collection skeleton (where galaxy_info['platforms']
makes less sense) in the near future. The galaxy['platforms'] variable was
never really used by Galaxy and is somewhat deprecated.

Closes #4.